Stop gracefully the ffmpeg process

+++ streaming/stream_manager.py
@@ -1,10 +1,10 @@
 import subprocess
 import os
 import time
+import signal
 from dotenv import load_dotenv
 
 load_dotenv()
-
 
 ffmpeg_process = None
 
@@ -35,7 +35,11 @@ def start_streaming(stop_event):
         f'{os.getenv("INGEST_URL")}{os.getenv("STREAM_KEY")}'
     ]
     
-    ffmpeg_process = subprocess.Popen(ffmpeg_command)
+    ffmpeg_process = subprocess.Popen(
+        ffmpeg_command,
+        stdin=subprocess.PIPE,
+        creationflags=subprocess.CREATE_NEW_PROCESS_GROUP
+    )
 
     try:
         print("🎥 Streaming del escritorio iniciado. Esperando mensaje de detención...")
@@ -49,12 +53,25 @@ def start_streaming(stop_event):
 def stop_streaming():
     global ffmpeg_process
     if ffmpeg_process:
-        print("🛑 Deteniendo el proceso de FFmpeg...")
-        ffmpeg_process.terminate()
-        ffmpeg_process.wait()
-        ffmpeg_process.kill()
-        ffmpeg_process = None
-        print("✅ Streaming finalizado.")
+        print("🛑 Intentando detener el proceso de FFmpeg...")
+        try:
+            if ffmpeg_process.poll() is None:
+                ffmpeg_process.send_signal(signal.CTRL_BREAK_EVENT)
+                print("🕒 Esperando que el proceso de FFmpeg termine...")
+                time.sleep(3)
+                if ffmpeg_process.poll() is None:
+                    print("⚠️ FFmpeg sigue activo. Intentando forzar la detención...")
+                    ffmpeg_process.terminate()
+                    time.sleep(1)
+                    if ffmpeg_process.poll() is None:
+                        print("🔴 Terminando proceso con fuerza...")
+                        ffmpeg_process.kill()
+            else:
+                print("✅ FFmpeg ya ha terminado.")
+        except Exception as e:
+            print(f"⚠️ Error al intentar detener FFmpeg: {e}")
+        finally:
+            ffmpeg_process = None
+            print("✅ Streaming finalizado.")
     else:
-        print("⚠️ No hay un proceso de streaming en ejecución")
-
+        print("⚠️ No hay un proceso de streaming en ejecución.")
